refactor: 카드 입력 스타일을 컴포넌트로 전환
