+++ app/src/style/CardStyles.ts
@@ -1,29 +1,5 @@
 import styled from "@emotion/styled";
 
-type ErrorFlag = {
-  isError?: boolean;
-};
-
-export const CardInput = styled.input<ErrorFlag>`
-  border: solid 1px ${(props) => (props.isError ? "#FF3D3D" : "#acacac")};
-  border-radius: 2px;
-  padding: 0.5rem;
-  font-size: 11px;
-  height: 44px;
-  width: 100%;
-  box-sizing: border-box;
-  -moz-appearance: textfield;
-  &::placeholder {
-    color: #acacac;
-  }
-
-  &::-webkit-inner-spin-button,
-  &::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-`;
-
 export const CardFieldset = styled.fieldset`
   border: none;
   display: flex;
